initial

+++ packages/service-authentication/internal/core/domain/user.go
@@ -1,11 +1,11 @@
 package domain
 
 type User struct {
-    Id int `json:"id"`
-    Uid string `json:"uid"`
-    EmailAddress string `json:"emailAddress"`
-    Password string `json:"password"`
-    FirstName string `json:"firstName"`
-    LastName string `json:"lastName"`
-    CreatedAt string `json:"createdAt"`
+	Id           int8   `json:"id"`
+	Uid          string `json:"uid"`
+	EmailAddress string `json:"emailAddress"`
+	Password     string `json:"password"`
+	FirstName    string `json:"firstName"`
+	LastName     string `json:"lastName"`
+	CreatedAt    string `json:"createdAt"`
 }
